refactor(wallet-info-screen): SingleSigInfoScreen, MultisigInfoScreen과 각각의 ViewModel을 WalletInfoScreen, WalletInfoViewModel로 통합 리팩토링 작업
