+++ lib/screens/wallet_info/single_sig_wallet_info_screen.dart
@@ -1,30 +1,20 @@
 import 'dart:async';
 
-import 'package:coconut_design_system/coconut_design_system.dart';
 import 'package:coconut_vault/constants/app_routes.dart';
 import 'package:coconut_vault/enums/pin_check_context_enum.dart';
 import 'package:coconut_vault/enums/wallet_enums.dart';
 import 'package:coconut_vault/localization/strings.g.dart';
 import 'package:coconut_vault/providers/auth_provider.dart';
-import 'package:coconut_vault/providers/view_model/wallet_info/single_sig_wallet_info_view_model.dart';
+import 'package:coconut_vault/providers/view_model/wallet_info/wallet_info_view_model.dart';
 import 'package:coconut_vault/providers/wallet_provider.dart';
-import 'package:coconut_vault/screens/home/select_sync_option_bottom_sheet.dart';
-import 'package:coconut_vault/screens/wallet_info/name_and_icon_edit_bottom_sheet.dart';
-import 'package:coconut_vault/utils/text_utils.dart';
-import 'package:coconut_vault/utils/vibration_util.dart';
-import 'package:coconut_vault/widgets/button/button_group.dart';
-import 'package:coconut_vault/widgets/button/single_button.dart';
-import 'package:coconut_vault/widgets/wallet_info/wallet_info_layout.dart';
+import 'package:coconut_vault/screens/wallet_info/wallet_info_layout.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:coconut_vault/screens/common/pin_check_screen.dart';
 import 'package:coconut_vault/widgets/bottom_sheet.dart';
-import 'package:coconut_vault/widgets/bubble_clipper.dart';
 import 'package:coconut_vault/widgets/custom_loading_overlay.dart';
 import 'package:provider/provider.dart';
-import 'package:shimmer/shimmer.dart';
 
-class SingleSigWalletInfoScreen extends StatefulWidget {
+class SingleSigWalletInfoScreen extends StatelessWidget {
   final int id;
   // 서명 전용 모드에서는 항상 false입니다.
   final bool shouldShowPassphraseVerifyMenu;
@@ -35,29 +25,6 @@ class SingleSigWalletInfoScreen extends StatefulWidget {
     required this.shouldShowPassphraseVerifyMenu,
     this.entryPoint,
   });
-
-  @override
-  State<SingleSigWalletInfoScreen> createState() => _SingleSigWalletInfoScreenState();
-}
-
-class _SingleSigWalletInfoScreenState extends State<SingleSigWalletInfoScreen> {
-  final GlobalKey _tooltipIconKey = GlobalKey();
-  RenderBox? _tooltipIconRendBox;
-  Offset _tooltipIconPosition = Offset.zero;
-  double _tooltipTopPadding = 0;
-
-  Timer? _tooltipTimer;
-  bool _isTooltipVisible = false;
-
-  @override
-  void initState() {
-    super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      _tooltipIconRendBox = _tooltipIconKey.currentContext?.findRenderObject() as RenderBox;
-      _tooltipIconPosition = _tooltipIconRendBox!.localToGlobal(Offset.zero);
-      _tooltipTopPadding = MediaQuery.paddingOf(context).top + kToolbarHeight - 14;
-    });
-  }
 
   Future<void> _authenticateWithBiometricOrPin(
     BuildContext context,
@@ -93,439 +60,68 @@ class _SingleSigWalletInfoScreenState extends State<SingleSigWalletInfoScreen> {
 
   @override
   Widget build(BuildContext context) {
-    return PopScope(
-      canPop: true,
-      onPopInvokedWithResult: (didPop, _) {
-        _removeTooltip();
-      },
-      child: CustomLoadingOverlay(
-        child: ChangeNotifierProvider(
-          create:
-              (context) => SingleSigWalletInfoViewModel(Provider.of<WalletProvider>(context, listen: false), widget.id),
-          child: Consumer<SingleSigWalletInfoViewModel>(
-            builder: (context, viewModel, child) {
-              final canDelete = viewModel.hasLinkedMultisigVault != true;
-              final walletName = viewModel.name;
-
-              return GestureDetector(
-                onTap: () => _removeTooltip(),
-                child: Scaffold(
-                  backgroundColor: CoconutColors.white,
-                  appBar: CoconutAppBar.build(
-                    title: walletName,
-                    context: context,
-                    // isBottom: viewModel.hasLinkedMultisigVault,
-                    isBottom: false,
-                    actionButtonList: [
-                      IconButton(
-                        onPressed: () {
-                          _removeTooltip();
-                          if (canDelete) {
-                            _showDeleteDialog(context, walletName);
-                            return;
-                          }
-                          CoconutToast.showToast(
-                            context: context,
-                            text: t.toast.name_multisig_in_use,
-                            isVisibleIcon: true,
-                          );
-                        },
-                        icon: SvgPicture.asset(
-                          'assets/svg/trash.svg',
-                          width: 20,
-                          colorFilter: ColorFilter.mode(
-                            canDelete ? CoconutColors.red : CoconutColors.gray850.withValues(alpha: 0.15),
-                            BlendMode.srcIn,
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                  body: WalletInfoLayout<SingleSigWalletInfoViewModel>(
-                    tooltipKey: _tooltipIconKey,
-                    onTooltipClicked: () => _toggleTooltipVisible(context),
-                    onNameChangeClicked: (viewModel) {
-                      _removeTooltip();
-                      _showModalBottomSheetForEditingNameAndIcon(viewModel);
-                    },
-                    linkedInfoBuilder:
-                        (viewModel) =>
-                            viewModel.hasLinkedMultisigVault == true
-                                ? _buildLinkedMultisigVaultInfoCard(context)
-                                : null,
-                    menuListBuilder: (viewModel) => _buildMenuList(context),
-                    tooltipBuilder: (viewModel) => _buildTooltip(context),
-                  ),
-                ),
-              );
-            },
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildMenuList(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16),
-      child: ButtonGroup(
-        buttons: [
-          SingleButton(
-            enableShrinkAnim: true,
+    return ChangeNotifierProvider(
+      create:
+          (context) => WalletInfoViewModel(Provider.of<WalletProvider>(context, listen: false), id, isMultisig: false),
+      child: WalletInfoLayout(
+        id: id,
+        shouldShowPassphraseVerifyMenu: shouldShowPassphraseVerifyMenu,
+        isMultisig: false,
+        menuButtonDatas: [
+          SingleButtonData(
             title: t.vault_menu_screen.title.single_sig_sign,
-            onPressed: () {
-              _removeTooltip();
-              Navigator.pushNamed(context, AppRoutes.psbtScanner, arguments: {'id': widget.id});
-            },
+            enableShrinkAnim: true,
+            onPressed: () => Navigator.pushNamed(context, AppRoutes.psbtScanner, arguments: {'id': id}),
           ),
-          SingleButton(
+          SingleButtonData(
             title: t.view_address,
             enableShrinkAnim: true,
-            onPressed: () {
-              _removeTooltip();
-              Navigator.pushNamed(
-                context,
-                AppRoutes.addressList,
-                arguments: {'id': widget.id, 'isSpecificVault': true},
-              );
-            },
+            onPressed:
+                () =>
+                    Navigator.pushNamed(context, AppRoutes.addressList, arguments: {'id': id, 'isSpecificVault': true}),
           ),
-          SingleButton(
+          SingleButtonData(
             title: t.vault_menu_screen.view_xpub,
             enableShrinkAnim: true,
-            onPressed: () {
-              _removeTooltip();
-              Navigator.pushNamed(context, AppRoutes.viewXpub, arguments: {'id': widget.id});
-            },
+            onPressed: () => Navigator.pushNamed(context, AppRoutes.viewXpub, arguments: {'id': id}),
           ),
-          SingleButton(
+          SingleButtonData(
             title: t.view_mnemonic,
             enableShrinkAnim: true,
             onPressed: () {
-              _removeTooltip();
-              if (!mounted) return;
-              final viewModel = context.read<SingleSigWalletInfoViewModel>();
-              if (viewModel.isSigningOnlyMode) {
-                Navigator.pushNamed(context, AppRoutes.mnemonicView, arguments: {'id': widget.id});
+              if (!context.mounted) return;
+              final walletProvider = context.read<WalletProvider>();
+              if (walletProvider.isSigningOnlyMode) {
+                Navigator.pushNamed(context, AppRoutes.mnemonicView, arguments: {'id': id});
                 return;
               }
 
               _authenticateWithBiometricOrPin(
                 context,
                 PinCheckContextEnum.sensitiveAction,
-                () => Navigator.pushNamed(context, AppRoutes.mnemonicView, arguments: {'id': widget.id}),
+                () => Navigator.pushNamed(context, AppRoutes.mnemonicView, arguments: {'id': id}),
               );
             },
           ),
-          if (widget.shouldShowPassphraseVerifyMenu) ...[
-            SingleButton(
-              enableShrinkAnim: true,
+          if (shouldShowPassphraseVerifyMenu) ...{
+            SingleButtonData(
               title: t.verify_passphrase,
-              onPressed: () {
-                _removeTooltip();
-                Navigator.of(context).pushNamed(AppRoutes.passphraseVerification, arguments: {'id': widget.id});
-              },
+              enableShrinkAnim: true,
+              onPressed: () => Navigator.pushNamed(context, AppRoutes.passphraseVerification, arguments: {'id': id}),
             ),
-          ],
-          SingleButton(
+          },
+          SingleButtonData(
             title: t.vault_menu_screen.export_wallet,
-            onPressed: () {
-              _removeTooltip();
-              Navigator.pushNamed(
-                context,
-                AppRoutes.vaultExportOptions,
-                arguments: {'id': widget.id, 'walletType': WalletType.singleSignature},
-              );
-            },
+            enableShrinkAnim: true,
+            onPressed:
+                () => Navigator.pushNamed(
+                  context,
+                  AppRoutes.vaultExportOptions,
+                  arguments: {'id': id, 'walletType': WalletType.singleSignature},
+                ),
           ),
-          // SingleButton(
-          //   title: t.vault_menu_screen.title.use_as_multisig_signer,
-          //   enableShrinkAnim: true,
-          //   onPressed: () {
-          //     _removeTooltip();
-          //     Navigator.pushNamed(context, AppRoutes.multisigSignerBsmsExport, arguments: {'id': widget.id});
-          //   },
-          // ),
-          // SingleButton(
-          //   enableShrinkAnim: true,
-          //   title: t.select_export_type_screen.title,
-          //   onPressed: () async {
-          //     if (widget.shouldShowPassphraseVerifyMenu) {
-          //       Seed? seed = await MyBottomSheet.showBottomSheet_ratio<Seed?>(
-          //         ratio: 0.5,
-          //         context: context,
-          //         child: PassphraseCheckScreen(id: widget.id, context: PassphraseCheckContext.export),
-          //       );
-
-          //       if (seed == null) {
-          //         return;
-          //       }
-
-          //       if (mounted) {
-          //         _showSyncOptionBottomSheet(widget.id, context);
-          //       }
-          //     } else {
-          //       _showSyncOptionBottomSheet(widget.id, context);
-          //     }
-          //   },
-          // ),
         ],
       ),
     );
-  }
-
-  void _showSyncOptionBottomSheet(int walletId, BuildContext context) {
-    MyBottomSheet.showBottomSheet_ratio(
-      context: context,
-      ratio: 0.5,
-      child: SelectSyncOptionBottomSheet(
-        onSyncOptionSelected: (format) {
-          if (!context.mounted) return;
-          Navigator.pop(context);
-          Navigator.pushNamed(context, AppRoutes.syncToWallet, arguments: {'id': walletId, 'syncOption': format});
-        },
-      ),
-    );
-  }
-
-  void _showModalBottomSheetForEditingNameAndIcon(SingleSigWalletInfoViewModel viewModel) {
-    MyBottomSheet.showBottomSheet_90(
-      context: context,
-      child: NameAndIconEditBottomSheet(
-        name: viewModel.name,
-        iconIndex: viewModel.iconIndex,
-        colorIndex: viewModel.colorIndex,
-        onUpdate: (String newName, int newIconIndex, int newColorIndex) {
-          setState(() {
-            _updateVaultInfo(newName, newColorIndex, newIconIndex, viewModel);
-          });
-        },
-      ),
-    );
-  }
-
-  void _updateVaultInfo(
-    String newName,
-    int newColorIndex,
-    int newIconIndex,
-    SingleSigWalletInfoViewModel viewModel,
-  ) async {
-    // 변경 사항이 없는 경우
-    if (newName == viewModel.name && newIconIndex == viewModel.iconIndex && newColorIndex == viewModel.colorIndex) {
-      return;
-    }
-
-    bool hasChanged = await viewModel.updateVault(widget.id, newName, newColorIndex, newIconIndex);
-
-    if (mounted) {
-      if (hasChanged) {
-        CoconutToast.showToast(context: context, text: t.toast.data_updated, isVisibleIcon: true);
-        return;
-      }
-      CoconutToast.showToast(context: context, text: t.toast.name_already_used, isVisibleIcon: true);
-    }
-  }
-
-  Widget _buildLinkedMultisigVaultInfoCard(BuildContext context) {
-    const linearGradient = LinearGradient(
-      colors: [Color(0xFFB2E774), Color(0xFF6373EB), Color(0xFF2ACEC3)],
-      begin: Alignment.centerLeft,
-      end: Alignment.centerRight,
-    );
-
-    return Container(
-      margin: const EdgeInsets.only(left: 16, right: 16),
-      decoration: BoxDecoration(
-        color: CoconutColors.white,
-        borderRadius: BorderRadius.circular(12),
-        gradient: linearGradient,
-      ),
-      child: Container(
-        margin: const EdgeInsets.all(1),
-        padding: const EdgeInsets.only(left: 20, right: 20, top: 20, bottom: 14),
-        decoration: BoxDecoration(color: CoconutColors.white, borderRadius: BorderRadius.circular(11)),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            // Icon & used in multisig vault text
-            Row(
-              children: [
-                SvgPicture.asset('assets/svg/vault-grey.svg', width: 18),
-                const SizedBox(width: 10),
-                Expanded(child: Text(t.vault_settings.used_in_multisig, style: CoconutTypography.body2_14)),
-              ],
-            ),
-            const Padding(padding: EdgeInsets.only(top: 4, bottom: 4, left: 28), child: Divider()),
-
-            // Linked multisig vaults
-            Consumer<SingleSigWalletInfoViewModel>(
-              builder: (context, viewModel, child) {
-                return ListView.builder(
-                  physics: const NeverScrollableScrollPhysics(),
-                  itemCount: viewModel.linkedMutlsigVaultCount,
-                  shrinkWrap: true,
-                  itemBuilder: (context, index) {
-                    final id = viewModel.linkedMultisigInfo!.keys.elementAt(index);
-                    final idx = viewModel.linkedMultisigInfo!.values.elementAt(index);
-
-                    if (viewModel.isLoadedVaultList && viewModel.existsLinkedMultisigVault(id)) {
-                      final multisig = viewModel.getVaultById(id);
-
-                      return InkWell(
-                        onTap: () {
-                          Navigator.pushNamed(context, AppRoutes.multisigSetupInfo, arguments: {'id': id});
-                        },
-                        child: Container(
-                          padding: const EdgeInsets.only(left: 28, bottom: 4),
-                          color: Colors.transparent,
-                          child: RichText(
-                            text: TextSpan(
-                              style: CoconutTypography.body2_14.setColor(const Color(0xFF4E83FF)),
-                              children: [
-                                TextSpan(
-                                  text: TextUtils.ellipsisIfLonger(multisig.name),
-                                  style: CoconutTypography.body2_14_Bold.setColor(const Color(0xFF4E83FF)),
-                                ),
-                                TextSpan(text: t.vault_settings.of),
-                                TextSpan(
-                                  text: t.vault_settings.nth(index: idx + 1),
-                                  style: CoconutTypography.body2_14_Bold.setColor(const Color(0xFF4E83FF)),
-                                ),
-                                TextSpan(text: t.vault_settings.key),
-                              ],
-                            ),
-                          ),
-                        ),
-                      );
-                    } else {
-                      return Container(
-                        padding: const EdgeInsets.only(left: 28, bottom: 4),
-                        child: Shimmer.fromColors(
-                          baseColor: CoconutColors.gray300,
-                          highlightColor: CoconutColors.gray150,
-                          child: Container(height: 17, width: double.maxFinite, color: CoconutColors.gray300),
-                        ),
-                      );
-                    }
-                  },
-                );
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Future<void> _deleteVault(BuildContext context) async {
-    if (!mounted) return;
-    final viewModel = context.read<SingleSigWalletInfoViewModel>();
-    viewModel.deleteVault();
-    vibrateLight();
-    if (widget.entryPoint != null && widget.entryPoint == AppRoutes.vaultList) {
-      Navigator.popUntil(context, (route) {
-        return route.settings.name == AppRoutes.vaultList;
-      });
-    } else {
-      Navigator.popUntil(context, (route) => route.isFirst);
-    }
-    return;
-  }
-
-  Widget _buildTooltip(BuildContext context) {
-    return Visibility(
-      visible: _isTooltipVisible,
-      child: Positioned(
-        top: _tooltipIconPosition.dy - _tooltipTopPadding,
-        right: MediaQuery.sizeOf(context).width - _tooltipIconPosition.dx - 48,
-        child: GestureDetector(
-          onTap: () => _removeTooltip(),
-          child: ClipPath(
-            clipper: RightTriangleBubbleClipper(),
-            child: Container(
-              constraints: BoxConstraints(maxWidth: MediaQuery.sizeOf(context).width - 40),
-              padding: const EdgeInsets.only(top: 25, left: 10, right: 10, bottom: 10),
-              color: CoconutColors.gray800,
-              child: Row(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Flexible(
-                    child: Text(
-                      t.tooltip.mfp,
-                      style: CoconutTypography.body3_12.merge(const TextStyle(height: 1.3, color: CoconutColors.white)),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  void _toggleTooltipVisible(BuildContext context) {
-    /// 이미 보여지고 있는 상태라면 툴팁 제거만 합니다.
-    if (_isTooltipVisible) {
-      _removeTooltip();
-      return;
-    }
-
-    setState(() {
-      _isTooltipVisible = true;
-    });
-
-    _tooltipTimer = Timer(const Duration(seconds: 5), () {
-      setState(() {
-        _isTooltipVisible = false;
-      });
-    });
-  }
-
-  void _removeTooltip() {
-    _tooltipTimer?.cancel();
-    setState(() {
-      _isTooltipVisible = false;
-    });
-  }
-
-  void _showDeleteDialog(BuildContext context, String walletName) {
-    showDialog(
-      context: context,
-      builder: (BuildContext dialogContext) {
-        return CoconutPopup(
-          insetPadding: EdgeInsets.symmetric(horizontal: MediaQuery.of(context).size.width * 0.15),
-          title: t.alert.delete_vault.title,
-          description: t.alert.delete_vault.description,
-          backgroundColor: CoconutColors.white,
-          leftButtonText: t.no,
-          leftButtonColor: CoconutColors.black.withValues(alpha: 0.7),
-          rightButtonText: t.yes,
-          rightButtonColor: CoconutColors.warningText,
-          onTapLeft: () => Navigator.pop(context),
-          onTapRight: () async {
-            if (!mounted) return;
-            final viewModel = context.read<SingleSigWalletInfoViewModel>();
-            if (!viewModel.isSigningOnlyMode) {
-              await _authenticateWithBiometricOrPin(
-                context,
-                PinCheckContextEnum.seedDeletion,
-                () => _deleteVault(context),
-              );
-            } else {
-              _deleteVault(context);
-            }
-          },
-        );
-      },
-    );
-
-    return;
-  }
-
-  @override
-  void dispose() {
-    _tooltipTimer?.cancel();
-    super.dispose();
   }
 }
