--- lib/widgets/wallet_info/wallet_info_layout.dart
@@ -1,63 +1,0 @@
-import 'package:coconut_design_system/coconut_design_system.dart';
-import 'package:coconut_vault/model/common/vault_list_item_base.dart';
-import 'package:coconut_vault/providers/view_model/wallet_info/vault_setup_info_view_model_base.dart';
-import 'package:coconut_vault/widgets/wallet_info/wallet_info_item_card.dart';
-import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-
-class WalletInfoLayout<T extends VaultSetupInfoViewModelBase<VaultListItemBase>> extends StatelessWidget {
-  final GlobalKey tooltipKey;
-  final VoidCallback onTooltipClicked;
-  final void Function(T viewModel) onNameChangeClicked;
-  final Widget? Function(T viewModel)? linkedInfoBuilder;
-  final Widget Function(T viewModel)? contentBuilder;
-  final Widget Function(T viewModel) menuListBuilder;
-  final Widget Function(T viewModel) tooltipBuilder;
-
-  const WalletInfoLayout({
-    super.key,
-    required this.tooltipKey,
-    required this.onTooltipClicked,
-    required this.onNameChangeClicked,
-    required this.menuListBuilder,
-    required this.tooltipBuilder,
-    this.linkedInfoBuilder,
-    this.contentBuilder,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Consumer<T>(
-      builder: (context, viewModel, child) {
-        return SafeArea(
-          child: SingleChildScrollView(
-            child: Stack(
-              children: [
-                Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    CoconutLayout.spacing_500h,
-                    WalletInfoItemCard<T>(
-                      tooltipKey: tooltipKey,
-                      onTooltipClicked: onTooltipClicked,
-                      onNameChangeClicked: onNameChangeClicked,
-                    ),
-                    if (linkedInfoBuilder != null) ...[
-                      CoconutLayout.spacing_300h,
-                      linkedInfoBuilder!(viewModel) ?? const SizedBox.shrink(),
-                    ],
-                    if (contentBuilder != null) contentBuilder!(viewModel),
-                    CoconutLayout.spacing_500h,
-                    menuListBuilder(viewModel),
-                    CoconutLayout.spacing_1500h,
-                  ],
-                ),
-                tooltipBuilder(viewModel),
-              ],
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
